Step4-2

+++ controllers/blogs.js
@@ -1,13 +1,15 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs)
 })
 
 blogsRouter.get('/:id', async (request, response) => {
-    const blog = await Blog.findById(request.params.id)
+    const blog = await Blog.findById(request.params.id).populate('user', { username: 1, name: 1 })
     if (blog) {
         response.json(blog)
     } else {
@@ -15,17 +17,24 @@ blogsRouter.get('/:id', async (request, response) => {
     }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
-    const blog = new Blog({...request.body, likes: request.body.likes || 0})
+blogsRouter.post('/', async (request, response, next) => {    
     try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+        const user = await User.findById(decodedToken.id)
+        const blog = new Blog({...request.body, likes: request.body.likes || 0, user: user._id})
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
         response.json(savedBlog)
     } catch (error) {
         next(error)
     }
 })
 
-blogsRouter.put('/:id', (request, response, next) => {
+blogsRouter.put('/:id', async (request, response, next) => {
     const body = request.body
   
     const blog = {
@@ -35,16 +44,30 @@ blogsRouter.put('/:id', (request, response, next) => {
         likes: body.likes
     }
   
-    Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
-        .then(updatedBlog => {
-            response.json(updatedBlog.toJSON())
-        })
-        .catch(error => next(error))
+    try {
+        const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
+        response.json(updatedBlog.toJSON())
+    } catch (error) {
+        next(error)
+    }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+    try {
+        const decodedToken = jwt.verify(request.token, process.env.SECRET)
+        if (!request.token || !decodedToken.id) {
+            return response.status(401).json({ error: 'token missing or invalid' })
+        }
+        const blog = await Blog.findById(request.params.id)
+        if(blog.user.toString() === decodedToken.id) {
+            await Blog.findByIdAndRemove(request.params.id)
+            response.status(204).end()
+        } else {
+            return response.status(401).json({ error: 'a blog can be deleted only by the user who added the blog' })
+        }        
+    } catch (error) {
+        next(error)
+    }
 })
 
 module.exports = blogsRouter
